+++ commands/statistics/definitions/cookies.js
